multi-state: do not compare to previous state
In iofunc.multi-state.js: When setting a MultiState, do not compare to the previous value as it might not reflect the true state of the MultiState. Instead, just set the state regardless. This solves issues with MultiStates where two distinct objects are used for a single state (e.g. in KNX installations where it is common to use a feedback object from the actor itself instead of relying on the state send by a switch).